Moved Exceptions and DTOs to a Common project for now.
