+++ src/API/Core/Winterhaven.API.Core.Application/Requests/Players/GetPlayer/GetPlayerRequestHandler.cs
@@ -6,10 +6,10 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using Winterhaven.API.Core.Application.Contexts.Users;
-using Winterhaven.API.Core.Application.Exceptions;
 using Winterhaven.API.Core.Application.Work.Users;
 using Winterhaven.API.Core.Domain.Entities.Players;
 using Winterhaven.API.Core.Domain.Entities.Users;
+using Winterhaven.Common.Exceptions;
 
 /// <summary>
 /// Provides a request handler used to fetch an existing <see cref="Player"/> from the current <see cref="UserAccount"/>.
