+++ src/API/Core/Winterhaven.API.Core.Application/Requests/Users/LoginUser/LoginUserRequestHandler.cs
@@ -5,12 +5,12 @@ using Microsoft.Extensions.Logging;
 using System;
 using System.Threading;
 using System.Threading.Tasks;
-using Winterhaven.API.Core.Application.Exceptions;
 using Winterhaven.API.Core.Application.Services.Security;
 using Winterhaven.API.Core.Application.Services.Users;
 using Winterhaven.API.Core.Application.Work;
 using Winterhaven.API.Core.Application.Work.Users;
 using Winterhaven.API.Core.Domain.Entities.Users;
+using Winterhaven.Common.Exceptions;
 
 /// <summary>
 /// Provides a request handler used to authenticate and enforce single-session login for a potential
