+++ src/API/Core/Winterhaven.API.Core.Application/Requests/Users/LogoutUser/LogoutUserRequestHandler.cs
@@ -6,11 +6,10 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using Winterhaven.API.Core.Application.Contexts.Users;
-using Winterhaven.API.Core.Application.Exceptions;
 using Winterhaven.API.Core.Application.Work;
 using Winterhaven.API.Core.Application.Work.Users;
 using Winterhaven.API.Core.Domain.Entities.Users;
-using Winterhaven.API.Core.Domain.Exceptions;
+using Winterhaven.Common.Exceptions;
 
 /// <summary>
 /// Provides a request handler used to logout a <see cref="UserAccount"/> and invalidate their
