+++ src/API/Core/Winterhaven.API.Core.Application/Winterhaven.API.Core.Application.csproj
@@ -5,6 +5,7 @@
     <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\..\..\Common\Winterhaven.Common\Winterhaven.Common.csproj" />
     <ProjectReference Include="..\Winterhaven.API.Core.Domain\Winterhaven.API.Core.Domain.csproj" />
   </ItemGroup>
     <ItemGroup>
