+++ src/API/Infrastructure/Winterhaven.API.Infrastructure/Contexts/ActorContext.cs
@@ -4,9 +4,9 @@ using Microsoft.AspNetCore.Http;
 using System;
 using System.Security.Claims;
 using Winterhaven.API.Core.Application.Contexts.Users;
-using Winterhaven.API.Core.Application.Exceptions;
-using Winterhaven.API.Core.Application.Work.Users;
+using Winterhaven.Common.Exceptions;
 using Winterhaven.API.Core.Domain.Entities.Users;
+using Winterhaven.API.Core.Application.Work.Users;
 
 internal sealed class ActorContext : IActorContext
 {
