+++ src/API/Presentation/Winterhaven.API.Presentation/Controllers/Players/PlayerController.cs
@@ -10,7 +10,7 @@ using System.Threading.Tasks;
 using Winterhaven.API.Core.Application.Requests.Players.GetPlayer;
 using Winterhaven.API.Core.Application.Requests.Players.UpdatePlayer;
 using Winterhaven.API.Core.Domain.Entities.Players;
-using Winterhaven.API.Presentation.DTOs.Players;
+using Winterhaven.API.Common.DTOs.Players;
 
 /// <summary>
 /// Provides API endpoints for managing <see cref="Player"/> entities.
