+++ src/API/Presentation/Winterhaven.API.Presentation/Mappings/Players/PlayerMapper.cs
@@ -4,7 +4,7 @@ using AutoMapper;
 using System.Diagnostics.CodeAnalysis;
 using Winterhaven.API.Core.Application.Requests.Players.UpdatePlayer;
 using Winterhaven.API.Core.Domain.Entities.Players;
-using Winterhaven.API.Presentation.DTOs.Players;
+using Winterhaven.API.Common.DTOs.Players;
 
 [ExcludeFromCodeCoverage]
 internal sealed class PlayerMapper : Profile
