+++ src/API/Presentation/Winterhaven.API.Presentation/Mappings/Users/UserMapper.cs
@@ -5,7 +5,7 @@ using System.Diagnostics.CodeAnalysis;
 using Winterhaven.API.Core.Application.Requests.Users.LoginUser;
 using Winterhaven.API.Core.Application.Requests.Users.RefreshToken;
 using Winterhaven.API.Core.Application.Requests.Users.RegisterUser;
-using Winterhaven.API.Presentation.DTOs.Users;
+using Winterhaven.API.Common.DTOs.Users;
 
 [ExcludeFromCodeCoverage]
 internal sealed class UserMapper : Profile
