+++ src/API/Tests/Winterhaven.API.Tests/Core/Application/Requests/Players/GetPlayer/GetPlayerRequestHandlerTests.cs
@@ -6,12 +6,12 @@ using NUnit.Framework;
 using System;
 using System.Threading.Tasks;
 using Winterhaven.API.Core.Application.Contexts.Users;
-using Winterhaven.API.Core.Application.Exceptions;
 using Winterhaven.API.Core.Application.Requests.Players.GetPlayer;
 using Winterhaven.API.Core.Application.Work.Users;
 using Winterhaven.API.Core.Domain.Entities.Players;
 using Winterhaven.API.Core.Domain.Entities.Users;
-using Winterhaven.API.Presentation.DTOs.Players;
+using Winterhaven.API.Common.DTOs.Players;
+using Winterhaven.Common.Exceptions;
 
 [TestFixture]
 internal sealed class GetPlayerRequestHandlerTests
