+++ src/API/Tests/Winterhaven.API.Tests/Infrastructure/Contexts/ActorContextTests.cs
@@ -6,10 +6,10 @@ using NSubstitute.ClearExtensions;
 using NUnit.Framework;
 using System;
 using System.Security.Claims;
-using Winterhaven.API.Core.Application.Exceptions;
 using Winterhaven.API.Core.Application.Work.Users;
 using Winterhaven.API.Core.Domain.Entities.Users;
 using Winterhaven.API.Infrastructure.Contexts;
+using Winterhaven.Common.Exceptions;
 
 [TestFixture]
 internal sealed class ActorContextTests
