+++ src/Common/Winterhaven.Common/DTOs/Maps/GetMapRequestDto.cs
@@ -1,4 +1,4 @@
-﻿namespace Winterhaven.API.Presentation.DTOs.Maps;
+﻿namespace Winterhaven.API.Common.DTOs.Maps;
 
 /// <summary>
 /// Represents the data transfer object used to fetch a map.
