+++ src/Common/Winterhaven.Common/DTOs/Players/GetPlayerRequestDto.cs
@@ -1,7 +1,6 @@
-﻿namespace Winterhaven.API.Presentation.DTOs.Players;
+﻿namespace Winterhaven.API.Common.DTOs.Players;
 
 using System.Diagnostics.CodeAnalysis;
-using Winterhaven.API.Core.Domain.Entities.Players;
 
 /// <summary>
 /// Represents a player within the system.
@@ -10,7 +9,8 @@ using Winterhaven.API.Core.Domain.Entities.Players;
 /// <param name="X">The current X-coordinate of the <see cref="Player"/>.</param>
 /// <param name="Y">The current Y-coordinate of the <see cref="Player"/>.</param>
 [ExcludeFromCodeCoverage]
-internal sealed record GetPlayerRequestDto(
+
+public sealed record GetPlayerRequestDto(
     string Name,
     double X,
     double Y);
